fix presentation format of DS and DNSKEY records
Algorithms were formatted as their enum name rather than their integer
value.  Convert them to `u8` (as is done with the digest type) and add
tests based on the examples in RFC 4034.

+++ crates/proto/src/rr/dnssec/rdata/dnskey.rs
@@ -445,7 +445,7 @@ impl fmt::Display for DNSKEY {
             f,
             "{flags} 3 {alg} {key}",
             flags = self.flags(),
-            alg = self.algorithm,
+            alg = u8::from(self.algorithm),
             key = data_encoding::BASE64.encode(&self.public_key)
         )
     }
